Add files via upload
This is a slightly different version that itterates through a range of IP addresses from one Initialization area of the BAT file. previously it was limited to 3 in file references to PI ip and Pi usernames.  Still a few known sleeping bugs I know of that seem to be working , such as GOTO within a parenthesis .